[Justice Counts] bug fix for records not being sent to FE for supervision subsystem data (Recidiviz/recidiviz-data#23644)
## Description of the change

This is a bug-fix for the incident that happened about two weeks ago
where a sheet was uploaded containing data for supervision subsystems
but the user saw `no metrics to review` when data was saved. I've
commented and given context around the fixes in in-line comments.

https://github.com/Recidiviz/recidiviz-data/assets/19961693/5e54e087-fd32-4b07-9a3c-676deb27cd8a

## Related issues

Closes Recidiviz/recidiviz-data#23305

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6956eb8240e9319037bf03498c0887d82b6db08e