[Justice Counts] Bulk Upload Updated Reports fix (Recidiviz/recidiviz-data#23715)
## Description of the change

Previously, if a user tried to upload monthly data for a metric that is
configured as annual (or vice versa) and the agency had previously
existing reports, we assumed that a previously existing reports existed
with the correct/new time range. This PR changes that assumption, so we
don't attempt to update reports statuses for reports that do not exist.

https://github.com/Recidiviz/recidiviz-data/assets/82831800/53f72be4-93b4-403b-a79e-8ec24cea9c4a

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: b49900a721d40ad020caeb627c6a24cda375be8f