Adds support for non-person-level event/span objects (Recidiviz/recidiviz-data#23567)
## Description of the change

This is the first of a few changes as part of [this
project](https://paper.dropbox.com/doc/WIP-Design-Doc-Extending-aggregated-metrics-infrastructure-to-usage-events--B~GPGvBlHvFtlCaQNKkkpx64Ag-O8wphbeCES6nTjzOekwmI)
to incorporate usage metrics and other metrics based on events and spans
at a unit of observation other than `PERSON`.

This change specifically converts `PersonEvent` and `PersonSpan` enums
into generic `Event` and `Span` enums, and makes the unit of analysis
type on the event or span a parameter of the object itself, rather than
a parameter that is passed into `generate_subquery`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#23511

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 50fe119be18497f2a4e7d11729a9b91f47d32e0f