[Justice Counts] Add all CSG users to older agencies (Recidiviz/recidiviz-data#23622)
## Description of the change

In late June, we merged the [following
PR](https://github.com/Recidiviz/recidiviz-data/pull/21773) which adds
all CSG users to newly created Agencies in production, and gives them
the READ_ONLY role.

We forgot to add all CSG users to previously created Agencies (agencies
created prior to that PR being merged/deployed).

This PR contains a script that adds all CSG users to all previously
created agencies. If the agency exists in production, we set their roles
to READ_ONLY. If the agency exists in staging, we set their roles to
AGENCY_ADMIN. Note that if a CSG user was previously assigned to a
particular agency with a particular role, we preserve that previous
relationship.

This script is intended to be run one time to account for older agencies
as CSG users should be added to newly created agencies moving forward.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23610

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: 4d12ae5754cf7bc18a475a6685fcabaea8d444b4