[US_IX] Fix state staff supervisor ingest to still grab employee info even if the employee is now inactive (Recidiviz/recidiviz-data#23761)
## Description of the change

Previously, the state staff supervisor views excluded employees that
were marked `Inactive` in the raw employee table. However, because we
receive the raw employee table every day but we only receive the
manually filled supervisor roster once a month, this was causing an
issue because supervisor periods were not getting ingested for employees
that were marked `Inactive` between the last and current roster dates.
And so supervisor periods for employees that left weren't getting
updated to be closed.

This PR fixes that by not excluding inactive employees, and instead
prioritizes grabbing the most recently active info for each staff
member.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0bc397b156f3e3c5d36c19d4e80c1e5b4279d4bc