[US_IX] Ingest officer role periods during supervisor periods ingest (Recidiviz/recidiviz-data#23767)
## Description of the change

Now that officer attribute sessions has been refactored to allow for
overlapping role periods, this PR revises supervisor periods ingest to
also ingest a role period with `role_subtype = 'SUPERVISION_OFFICER'`
for every officer we ingest a supervisor period for. We were already
ingesting a role period for every supervisor we ingest a supervisor
period for.

This PR also removes the workarounds that were put in to include Idaho
staff that held dual roles in the outliers officers and supervisors
views.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ea4931273d4515a961a784ff5253d6d66f4b23bf