[Airflow] Automate adding `.py` dependencies to source files to upload to bucket (Recidiviz/recidiviz-data#23482)
## Description of the change

Automate adding `.py` dependencies to source files to upload to the
bucket. The old method involved manually tracing file dependencies and
trying to figure out which ones should be added to
`cloud_composer_source_files_to_copy.yaml`. The new method reduces the
number of files uploaded to the bucket since it is more precise.

It doesn't currently remove files already in the bucket, but that issue
should be reduced with @ohaibbq's work on spinning up experimental dags.
Also, it doesn't handle non-python files, so Yaml and JSON files still
need to be listed manually.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#23184

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ccece603ea7784fe7e55955a07c6ee474458fc26