[Cleanup] Delete FakeDirectIngestInstanceStatusManager (Recidiviz/recidiviz-data#23803)
## Description of the change

As I started to investigate Recidiviz/recidiviz-data#23799, I discovered that there is a fake
implementation of `DirectIngestInstanceStatusManager` which we're using
in a number of tests because those tests use `freeze_time`, which messes
with our Postgres database invariants (can't have two statuses with the
same timestamp).

The logic in this fake class diverges from the actual logic we use to
commit ingest statuses so any tests that use it aren't really testing
what happens in real life - not ideal.

This PR:
a) Updates controller tests to stop using freeze_time and instead,
updates the actual production parser code to arbitrarily pick a
"from_dt" so that we can deterministically produce dates from duration
strings.
b) Updates all tests to just use
`PostgresDirectIngestInstanceStatusManager`
c) Deletes the `FakeDirectIngestInstanceStatusManager`.

This will be important going forward as we modify some of the status
manager logic to handle new post-ingest in Dataflow statuses. It also
will remove the requirement that time is frozen in
ingest-in-Dataflow-based pipeline integration tests.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#23799

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3dbb41c7d7508ef630962d8571b5b11de7d5378b