[Ingest in Dataflow] Limit scope of ingest_view_rank_list() (Recidiviz/recidiviz-data#23812)
## Description of the change

There are a handful of places we're instantiating an ingest controller
just to get the `ingest_view_rank_list`, where we really don't need to
be. This function is going to go away post ingest in Dataflow, so this
updates to eliminate a few places where we use it.

Does not address all usages, e.g. there are a bunch of places where we
still reference the rank list in controller tests, which we will
eliminate once we rewrite those tests to use the Dataflow pipeline under
the hood.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Will make some of the test fixes for Recidiviz/recidiviz-data#23799 easier.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e242cb193ba36a044185452685bf2626c6a9aa59