[US_IX] Hydrate admission/release reasons for housing unit changes (Recidiviz/recidiviz-data#23874)
## Description of the change

[This](https://github.com/Recidiviz/recidiviz-data/pull/22970) PR
revised incarceration periods by breaking up periods to be more granular
if housing unit / custody level changed within the period. That PR
accidentally left admission/release reason NULL in the cases where a
period was added to account for a housing unit change. This is leading
to zero-day periods not getting dropped in normalization, so this PR
hydrates admission/release reasons for these cases.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ccbbfcfedc1ce03747e4e661c3229d6b4dc1788f