[US_TN] Classification opp record query updates (Recidiviz/recidiviz-data#23860)
## Description of the change

This PR makes a couple updates to the opportunity record template for TN
CLD and Annual Reclass:
1) TTs had requested being able to see Incident Details along with the
Class and date of an incident. I'm not totally sure how this will look
on the front end, especially for Q6 and Q7 where we don't have a large
amount of space. Are there any updates we should do to allow wrap-text,
etc? (@Catacola @danielsmc )
2) Added in information about Incompatibles and Health Classification
needed for the summary sheet

Sandbox in `dsharm20230911_workflows_views`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#23800
Closes Recidiviz/recidiviz-data#23085

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 940ee4ed7bbc340517e2e00ae789015031449bac